- Panel III done - Start Date - End Date added (JdateChooser) -Layout Done -Close Button added - Panel IV Next

+++ src/Create_File.java
@@ -13,37 +13,61 @@ public class Create_File {
 
     public void create_File(String content) {
 
-        String path = "/home/hbrtxito/Downloads/hi.txt";
 
         String headline = "             B U D G E T   A P P L I CA T I O N  \n" +
                 "==========================================================================================="+
-                        "\n     Database's Back Up - Tex File Version - V.01 \n " +
+                "\n     Database's Back Up - Tex File Version - V.01 \n " +
                 "\n Quantity  -  Product    - Category  -   Store    -  Price   -   Payment  -  Date  - Comments \n" ;
 
 
-        try {
 
-            // create new file
-            File file = new File(path);
+        String username = System.getProperty("user.home");
+        String new_dir = "/media/hbrtxito/My Data/Budget_back_up";
+        System.out.println(username);
 
-            // if file doesn't exists, then create it
-            if (!file.exists() ) {
-                //file.createNewFile();
-                System.out.println("file created");
-                FileWriter writer = new FileWriter(file);
+
+        File directory = new File(new_dir+File.separator+"Budget_app_data");
+        String name_of_file = "budget_back_up.txt";
+
+        File file = new File(directory+File.separator+ name_of_file);
+
+        if (!directory.exists()){
+            try {
+                directory.mkdirs();
+
+
+                // if file doesn't exists, then create it
+                if (!file.exists() ) {
+                    //file.createNewFile();
+                    System.out.println("file created");
+                    FileWriter writer = new FileWriter(file);
+                    BufferedWriter bw = new BufferedWriter(writer);
+                    bw.write(headline);
+                    bw.close();
+
+                }
+                FileWriter writer = new FileWriter(file, true);
                 BufferedWriter bw = new BufferedWriter(writer);
-                bw.write(headline);
+                bw.write("\n"+ content);
                 bw.close();
-
+            }catch(IOException ex){
+                ex.printStackTrace();
             }
-            FileWriter writer = new FileWriter(path, true);
-            BufferedWriter bw = new BufferedWriter(writer);
-            bw.write("\n"+ content);
-            bw.close();
-        }catch(IOException ex){
+
+        }else {
+            System.out.println("The directory is already on the computer");
+            try {
+                FileWriter writer = new FileWriter(file, true);
+                BufferedWriter bw = new BufferedWriter(writer);
+                bw.write("\n"+ content);
+                bw.close();
+            }catch(IOException ex){
             ex.printStackTrace();
         }
 
+
+
+        }
     }
